requirements.txt
# LLM Evaluation - Korean Benchmark Tool
# Python 3.10+ required

# Core ML/LLM frameworks
vllm>=0.13.0
torch>=2.9.0
transformers>=4.56.0

# OpenAI API
openai>=2.14.0

# Data handling
datasets>=3.6.0
huggingface-hub>=0.36.0
numpy>=2.2.0

# Utilities
python-dotenv>=1.2.0
tqdm>=4.67.0
